[travelmux] NBD: breaking code into modules without changing behavior
