+++ services/travelmux/src/api/v4/mod.rs
@@ -1,1 +1,4 @@
+mod error;
 pub mod plan;
+
+pub use plan::Plan;
